pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pooh-code"
version = "0.1.0"
description = "Python reimplementation of the core Claude Code workflow with ChatGPT OAuth and Feishu WS access."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "duckduckgo-search>=8.1.1",
    "httpx>=0.28.1",
    "lark-oapi>=1.5.3",
    "oauth-cli-kit>=0.1.3",
    "python-dotenv>=1.2.2",
]

[project.scripts]
pooh-code = "pooh_code.main:main"
pooh-code-login = "pooh_code.openai_codex:login_main"

[tool.hatch.build.targets.wheel]
packages = ["src/pooh_code"]
